Start parsing operands
Operands will be parsed first, and then the options that are passed in.
The operands that dd uses are contained in a struct called Operands. The
operands will be parsed and put into the struct, so it can be used
easily later to determine which operations to perform.
Operands such as bs, obs, ibs, have parsing implemented.